+++ tools/tiff2rgba.c
@@ -203,7 +203,7 @@ static int cvt_by_tile(TIFF *in, TIFF *out)
     /*
      * Allocate tile buffer
      */
-    rastersize = tile_width * tile_height * sizeof(uint32_t);
+    rastersize = (uint32_t)(tile_width * tile_height * sizeof(uint32_t));
     if (tile_width != (rastersize / tile_height) / sizeof(uint32_t))
     {
         TIFFError(TIFFFileName(in),
@@ -221,7 +221,7 @@ static int cvt_by_tile(TIFF *in, TIFF *out)
      * Allocate a scanline buffer for swapping during the vertical
      * mirroring pass.
      */
-    wrk_linesize = tile_width * sizeof(uint32_t);
+    wrk_linesize = (uint32_t)(tile_width * sizeof(uint32_t));
     if (tile_width != wrk_linesize / sizeof(uint32_t))
     {
         TIFFError(TIFFFileName(in),
@@ -319,7 +319,7 @@ static int cvt_by_strip(TIFF *in, TIFF *out)
     /*
      * Allocate strip buffer
      */
-    rastersize = width * rowsperstrip * sizeof(uint32_t);
+    rastersize = (uint32_t)(width * rowsperstrip * sizeof(uint32_t));
     if (width != (rastersize / rowsperstrip) / sizeof(uint32_t))
     {
         TIFFError(TIFFFileName(in),
@@ -337,7 +337,7 @@ static int cvt_by_strip(TIFF *in, TIFF *out)
      * Allocate a scanline buffer for swapping during the vertical
      * mirroring pass.
      */
-    wrk_linesize = width * sizeof(uint32_t);
+    wrk_linesize = (uint32_t)(width * sizeof(uint32_t));
     if (width != wrk_linesize / sizeof(uint32_t))
     {
         TIFFError(TIFFFileName(in),
